test: combine all tests into a single class
