+++ lib/src/main/kotlin/team8/bstlib/AVLNode.kt
@@ -1,18 +1,14 @@
 package team8.bstlib
 
 class AVLNode<K : Comparable<K>, V>(key: K, value: V) : BinarySearchTreeNode<K, V, AVLNode<K, V>>(key, value) {
-    private var height: Int = 1
 
-    fun getHeight(): Int = height
-    // Обновляет высоту узла на основе высот потомков
+    var nodeHeight: Int = 1
+
     fun updateHeight() {
-        height = 1 + maxOf(
-            leftChild?.height ?: 0,
-            rightChild?.height ?: 0
-        )
+        val left = leftChild?.nodeHeight ?: 0
+        val right = rightChild?.nodeHeight ?: 0
+        nodeHeight = maxOf(left, right) + 1
     }
 
-    // Вычисляет баланс-фактор узла
-    fun balanceFactor() =
-        (leftChild?.height ?: 0) - (rightChild?.height ?: 0)
+    fun balanceFactor() = (leftChild?.nodeHeight ?: 0) - (rightChild?.nodeHeight ?: 0)
 }
